process-analysis-agent/requirements.txt
# requirements.txt
# Process Diagram Multi-Agent System Dependencies

# =============================================================================
# Core: Google ADK & Gemini
# =============================================================================
google-adk>=1.18.0
google-genai>=1.0.0
google-generativeai>=0.8.0

# =============================================================================
# Data Validation & Parsing (WICHTIG: Das fehlte!)
# =============================================================================
pydantic>=2.0.0     # Wird von fast allen deinen Agenten genutzt
PyPDF2>=3.0.0       # Wird von tools/pdf_parser.py genutzt

# =============================================================================
# Environment Management
# =============================================================================
python-dotenv==1.0.0

# =============================================================================
# Optional: Better Logging & Utilities
# =============================================================================
colorlog==6.8.0
reportlab           # Hilfreich, um Test-PDFs zu generieren (wie vorhin gemacht)

# =============================================================================
# Development
# =============================================================================
pytest>=7.4.0
pytest-asyncio>=0.21.0